requirements.txt
google-generativeai
python-dotenv
requests
textblob
translate
nltk